+++ src/final2/evolution_simulator/croisement.h
@@ -15,6 +15,5 @@ float croisement_continue(float trait1, float trait2, float coef);
 // ind1, ind2: les parents, 
 //  n_trs: nombre de traits
 // gen: génération du nouvel individu
-Individu* croisement(const Individu* ind1,const Individu* ind2, size_t n_trs, int gen);
-
+Individu* croisement(const Individu* ind1,const Individu* ind2, size_t n_trs, size_t gen);
 #endif // CROISEMENT_H_INCLUDED
