+++ src/final2/evolution_simulator/survie.h
@@ -18,5 +18,5 @@ float survie_continue(float c_ind, float c_env);
 // env: l'environnement
 // gen: la génération de l'individu
 // retourne un booléen (0: survie, 1: meurs)
-int survie_globale(const Individu* ind, const Destraits* trs, size_t n_trs, const Environnement* env, int gen);
+int survie_globale(const Individu* ind, const Destraits* trs, size_t n_trs, const Environnement* env, size_t gen);
 #endif
